third_party/blink/web_tests/platform/mac/external/wpt/css/css-text/white-space/white-space-collapse-002-expected.txt

This is a testharness.js-based test.
[FAIL] RLO 2
  assert_equals: expected 105 but got 112
[FAIL] RLO 3
  assert_equals: expected 105 but got 112
[FAIL] RLO 4
  assert_equals: expected 105 but got 112
[FAIL] RLO 5
  assert_equals: expected 105 but got 112
[FAIL] RLE 2
  assert_equals: expected 103 but got 109
[FAIL] RLE 3
  assert_equals: expected 103 but got 109
[FAIL] RLE 4
  assert_equals: expected 103 but got 109
[FAIL] RLE 5
  assert_equals: expected 103 but got 109
[FAIL] RLI 2
  assert_equals: expected 93 but got 100
[FAIL] RLI 3
  assert_equals: expected 93 but got 100
[FAIL] RLI 4
  assert_equals: expected 93 but got 100
[FAIL] RLI 5
  assert_equals: expected 93 but got 100
[FAIL] RLM 2
  assert_equals: expected 111 but got 117
[FAIL] RLM 3
  assert_equals: expected 111 but got 117
[FAIL] RLM 4
  assert_equals: expected 111 but got 117
[FAIL] RLM 5
  assert_equals: expected 111 but got 117
Harness: the test ran to completion.
